requirements.txt
djangorestframework
django-cors-headers
scikit-learn
matplotlib




